Transfer network util functions in appropriate class
