+++ src/Utils/NetworkUtils.cs
@@ -1,7 +1,9 @@
-﻿using System;
+﻿using RPVoiceChat.Networking;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Net;
+using System.Net.Http;
 
 namespace RPVoiceChat.Utils
 {
@@ -39,6 +41,47 @@ namespace RPVoiceChat.Utils
             catch (FormatException) { }
 
             throw new ArgumentException($"IP string \"{ipString}\" is not a valid IP address");
+        }
+
+        public static bool IsInternalNetwork(string ip)
+        {
+            return IsInternalNetwork(IPAddress.Parse(ip));
+        }
+
+        public static bool IsInternalNetwork(IPAddress ip)
+        {
+            byte[] ipParts = ip.GetAddressBytes();
+
+            if (ipParts[0] == 10 ||
+               (ipParts[0] == 192 && ipParts[1] == 168) ||
+               (ipParts[0] == 172 && (ipParts[1] >= 16 && ipParts[1] <= 31)) ||
+               (ipParts[0] == 25 || ipParts[0] == 26) ||
+               (ipParts[0] == 127 && ipParts[1] == 0 && ipParts[2] == 0 && ipParts[3] == 1))
+                return true;
+
+            return false;
+        }
+
+        public static IPEndPoint GetEndPoint(ConnectionInfo connectionInfo)
+        {
+            var address = IPAddress.Parse(connectionInfo.Address);
+            var endpoint = new IPEndPoint(address, connectionInfo.Port);
+
+            return endpoint;
+        }
+
+        public static string GetPublicIP()
+        {
+            string publicIPString = new HttpClient().GetStringAsync("https://ipinfo.io/ip").GetAwaiter().GetResult();
+
+            return publicIPString;
+        }
+
+        public static bool AssertEqual(IPEndPoint firstEndPoint, IPEndPoint secondEndPoint)
+        {
+            bool isSameAddress = firstEndPoint.Address.MapToIPv4().ToString() == secondEndPoint.Address.MapToIPv4().ToString();
+            bool isSamePort = firstEndPoint.Port == secondEndPoint.Port;
+            return isSameAddress && isSamePort;
         }
     }
 }
